First commit with all tests green for client
Removed ClientMethods due to no need for it has been seen

Removed getter and setter for portnumber since you connect and
disconnect with it. Might reimplement later if need arises.

Updatet tests and removed Mockito code which only hindered the process.

Added username and runnable to client to support multiples.